todo on imports, updates to validation files in routs and controllers
