+++ backend/routes/api/users.js
@@ -1,63 +1,24 @@
-// This file handles user CRUD: registration, information, edits, and deletion.
+// This file handles user routes
+// CRUD: registration, information, edits, and deletion logic is handled in the controllers
 // Authentication and authorization are handled in the routes/api/auth.js file.
 const express = require('express');
 const router = express.Router();
-const { check } = require('express-validator');
-const bcrypt = require('bcrypt');
-const { User } = db;
-const { Op } = require('sequelize');
+
 const { setTokenCookie, restoreUser } = require('../../utilities/auth');
-const { handleValidationErrors } = require('../utils/validation');
-
-
-const validateSignup = [
-    check('email')
-        .isLength({ min: 5 })
-        .withMessage('Email must be at least 5 characters'),
-    check('email')
-        .isLength({ max: 255 })
-        .withMessage('Email must be 255 characters or less'),
-    check('email')
-        .isEmail()
-        .withMessage('Invalid email'),
-    check('username')
-        .exists({ checkFalsy: true })
-        .withMessage('Username is required'),
-    check('username')
-        .not()
-        .isEmail()
-        .withMessage('Username cannot be an email'),
-    check('username')
-        .isLength({ min: 2 })
-        .withMessage('Username must be at least 2 characters'),
-    check('username')
-        .isLength({ max: 50 })
-        .withMessage('Username must be 50 characters or less'),
-    check('password')
-        .exists({ checkFalsy: true })
-        .withMessage('Password is required'),
-    handleValidationErrors
-];
-
+const { handleValidationErrors, } = require('../utils/validation');
+const { signup } = require('../../controllers/utils/userController');
+const { validateSignup } = require('../utils/userValidations');
 
 // Sign-up new User
-router.post('/', validateSignup);
-
+router.post('/', validateSignup, signup);
 
 // Get current user
-router.get('/', async (req, res, next) => {
-
-})
+router.get('/', )
 
 // Edit current user
-router.put('/', async (req, res, next) => {
-    
-})
+router.put('/', )
 
 // Delete current user
-router.delete('/', async (req, res, next) => {
-    
-})
-
+router.delete('/', )
 
 module.exports = router;
